# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Efrica website  OpenApi  Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "khumbozimba4@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1"
        }
    ],
    "paths": {
        "/article/getAllArticles": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Get list of articles",
                "description": "Returns list of articles",
                "operationId": "getAllArticles",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ArticleResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/article/addArticle": {
            "post": {
                "tags": [
                    "Articles"
                ],
                "summary": "Store new article",
                "description": "Returns article data",
                "operationId": "storeArticle",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreArticleRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Article"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/article/{id}/getArticleById": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Get article information",
                "description": "Returns article data",
                "operationId": "getArticleById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Article id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Article"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/article/{id}/updateArticleById": {
            "put": {
                "tags": [
                    "Articles"
                ],
                "summary": "Update existing article",
                "description": "Returns updated article data",
                "operationId": "updateArticle",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Article id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateArticleRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Article"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/article/{id}/deleteArticleById": {
            "delete": {
                "tags": [
                    "Articles"
                ],
                "summary": "Delete existing article",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteArticle",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Article id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Article": {
                "title": "Article",
                "description": "Article model",
                "properties": {
                    "title": {
                        "title": "title",
                        "description": "title of the Article model",
                        "example": ""
                    },
                    "photo_url": {
                        "title": "photo_url",
                        "description": "photo_url of the Article model",
                        "example": ""
                    },
                    "content": {
                        "title": "content",
                        "description": "content of the Article model",
                        "example": ""
                    },
                    "published_at": {
                        "title": "published_at",
                        "description": "published_at of the Article model",
                        "example": ""
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Article"
                }
            },
...
            "ArticleResource": {
                "title": "ArticleResource",
                "description": "Article resource",
                "properties": {
                    "data": {
                        "title": "Data",
                        "description": "Data wrapper",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Article"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "ArticleResource"
                }
            },
...
            "StoreArticleRequest": {
                "title": "Store Article request",
                "description": "Store Article request body data",
                "required": [
                    "title",
                    "photo_url",
                    "content",
                    "published_at"
                ],
                "properties": {
                    "title": {
                        "title": "title",
                        "description": "title of the new Article",
                        "type": "string",
                        "example": "Your example value here"
                    },
                    "photo_url": {
                        "title": "photo_url",
                        "description": "photo_url of the new Article",
                        "type": "string",
                        "example": "Your example value here"
                    },
                    "content": {
                        "title": "content",
                        "description": "content of the new Article",
                        "type": "string",
                        "example": "Your example value here"
                    },
                    "published_at": {
                        "title": "published_at",
                        "description": "published_at of the new Article",
                        "type": "string",
                        "example": "Your example value here"
                    }
                },
                "type": "object"
            },
...
            "UpdateArticleRequest": {
                "title": "UpdateArticle request",
                "description": "UpdateArticle request body data",
                "required": [
                    "title",
                    "photo_url",
                    "content",
                    "published_at"
                ],
                "properties": {
                    "title": {
                        "title": "title",
                        "description": "title of the new Article",
                        "type": "string",
                        "example": "Your example value here"
                    },
                    "photo_url": {
                        "title": "photo_url",
                        "description": "photo_url of the new Article",
                        "type": "string",
                        "example": "Your example value here"
                    },
                    "content": {
                        "title": "content",
                        "description": "content of the new Article",
                        "type": "string",
                        "example": "Your example value here"
                    },
                    "published_at": {
                        "title": "published_at",
                        "description": "published_at of the new Article",
                        "type": "string",
                        "example": "Your example value here"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}